requirements.txt
# Core dependencies
crewai~=0.102.0
langchain
litellm

# Smart contract analysis tools
slither-analyzer
py-solc-x
mythril
slither
# Machine learning and classification
transformers~=4.49.0
composio-core

# Environment configuration
python-dotenv~=1.0.1

pip~=24.3.1
configparser~=7.2.0
attrs~=25.1.0
distro~=1.9.0
wheel~=0.45.1
cryptography~=44.0.2
pillow~=11.1.0
filelock~=3.17.0
pytest~=8.3.5
setuptools~=75.6.0
packaging~=24.2
tomli~=2.2.1
autocommand~=2.2.2
inflect~=7.3.1
typeguard~=4.3.0
platformdirs~=4.3.6
zipp~=3.21.0
build~=1.2.2.post1
install~=1.3.4
pygments~=2.19.1
requests~=2.32.3
pandas~=2.2.3
beautifulsoup4~=4.13.3
openai~=1.65.3
numpy~=1.24.3
torch~=2.0.1
sklearn~=0.0.post7
scikit-learn~=1.3.0